refactor semantics
